Add tabbed customer detail view with modular lists
