+++ frontend/src/pages/CustomerDetail.jsx
@@ -1,8 +1,10 @@
 import { useEffect, useState } from 'react';
 import { useParams, useNavigate } from 'react-router-dom';
 import customerService from '../services/customerService';
-import reportService from '../services/reportService';
-import { ArrowLeft, FileText, AlertTriangle, DollarSign, CheckCircle, Calendar } from 'lucide-react';
+import { ArrowLeft, AlertTriangle } from 'lucide-react';
+import InvoiceList from '../components/customer/InvoiceList';
+import PaymentList from '../components/customer/PaymentList';
+import { formatCurrency } from '../utils/formatCurrency';
 
 const CustomerDetail = () => {
   const { id } = useParams();
@@ -10,11 +12,8 @@ const CustomerDetail = () => {
   const [customer, setCustomer] = useState(null);
   const [loading, setLoading] = useState(true);
   const [error, setError] = useState(null);
-  const [invoices, setInvoices] = useState([]);
-  const [payments, setPayments] = useState([]);
-  const [loadingInvoices, setLoadingInvoices] = useState(false);
-  const [loadingPayments, setLoadingPayments] = useState(false);
   const [summary, setSummary] = useState(null);
+  const [activeTab, setActiveTab] = useState('invoices');
 
   useEffect(() => {
     const fetchCustomer = async () => {
@@ -31,161 +30,8 @@ const CustomerDetail = () => {
     fetchCustomer();
   }, [id]);
 
-  // Tüm faturaları yükle
-  useEffect(() => {
-    const fetchAllInvoices = async () => {
-      if (!customer) return;
-      
-      try {
-        setLoadingInvoices(true);
-        
-        // Ödenmemiş faturaları al
-        const unpaidRes = await reportService.getCustomerUnpaidInvoicesSummary(id);
-        const unpaidData = unpaidRes.data.data;
-        
-        // Ödenmiş faturaları al
-        const paidRes = await reportService.getCustomerPaidInvoicesSummary(id);
-        const paidData = paidRes.data.data;
-        
-        // Faturaları birleştir
-        const allInvoices = [
-          ...unpaidData.invoices.map(invoice => ({
-            ...invoice,
-            status: 'unpaid',
-            isOverdue: invoice.isOverdue || false,
-            overdueDays: invoice.overdueDays || 0
-          })),
-          ...paidData.payments.map(invoice => ({
-            ...invoice,
-            status: 'paid',
-            lastPaymentDate: invoice.lastPaymentDate,
-            paymentMethod: invoice.paymentMethod
-          }))
-        ];
-        
-        // Tarihe göre sırala (en yeni önce)
-        allInvoices.sort((a, b) => new Date(b.date) - new Date(a.date));
-        
-        setInvoices(allInvoices);
-        
-        // Özet bilgileri
-        setSummary({
-          totalInvoices: allInvoices.length,
-          unpaidInvoices: unpaidData.invoices.length,
-          paidInvoices: paidData.payments.length,
-          totalUnpaidAmount: unpaidData.summary.totalAmount,
-          totalPaidAmount: paidData.summary.totalAmount,
-          overdueInvoices: unpaidData.summary.overdueInvoices,
-          overdueAmount: unpaidData.summary.overdueAmount,
-          lastPaymentDate: paidData.summary.lastPaymentDate
-        });
-        
-      } catch (err) {
-        console.error('Faturalar yüklenirken hata:', err);
-      } finally {
-        setLoadingInvoices(false);
-      }
-    };
-    fetchAllInvoices();
-  }, [id, customer]);
-
-  // Ödemeleri yükle
-  useEffect(() => {
-    const fetchPayments = async () => {
-      if (!customer) return;
-      
-      try {
-        setLoadingPayments(true);
-        const res = await reportService.getCustomerPayments(id);
-        setPayments(res.data.data.payments);
-      } catch (err) {
-        console.error('Ödemeler yüklenirken hata:', err);
-      } finally {
-        setLoadingPayments(false);
-      }
-    };
-    fetchPayments();
-  }, [id, customer]);
-
-  const formatCurrency = (amount) => {
-    return new Intl.NumberFormat('tr-TR', {
-      style: 'currency',
-      currency: 'TRY',
-    }).format(amount);
-  };
-
   const formatDate = (dateString) => {
     return new Date(dateString).toLocaleDateString('tr-TR');
-  };
-
-  const calculateDaysBetween = (startDateString, endDateInput) => {
-    const start = new Date(startDateString).getTime();
-    const end = endDateInput instanceof Date ? endDateInput.getTime() : new Date(endDateInput).getTime();
-    const diffMs = Math.max(0, end - start);
-    return Math.floor(diffMs / (1000 * 60 * 60 * 24));
-  };
-
-  const getElapsedDays = (invoice) => {
-    // Ödenmiş faturalar için: fatura tarihinden son ödeme tarihine kadar
-    // Ödenmemiş faturalar için: fatura tarihinden bugüne kadar
-    const endDate = invoice.status === 'paid' && invoice.lastPaymentDate
-      ? invoice.lastPaymentDate
-      : new Date();
-    return calculateDaysBetween(invoice.date, endDate);
-  };
-
-  const getStatusBadge = (invoice) => {
-    // Belge türüne göre durum belirleme
-    if (invoice.documentType) {
-      const docType = invoice.documentType.toLowerCase();
-      
-      // İade faturası kontrolü
-      if (docType.includes('iade') || docType.includes('return') || docType.includes('red')) {
-        return <span className="badge badge-info">İade</span>;
-      }
-      
-      // Devir fişi kontrolü
-      if (docType.includes('devir') && docType.includes('fiş')) {
-        if (invoice.status === 'paid') {
-          return <span className="badge badge-success">Devir Fişi (Ödenmiş)</span>;
-        } else if (invoice.isOverdue) {
-          return <span className="badge badge-error">Devir Fişi ({invoice.overdueDays} gün gecikmiş)</span>;
-        } else {
-          return <span className="badge badge-warning">Devir Fişi (Ödenmemiş)</span>;
-        }
-      }
-      
-      // Satış faturası kontrolü
-      if (docType.includes('satış') || docType.includes('sales')) {
-        if (invoice.status === 'paid') {
-          return <span className="badge badge-success">Satış Faturası (Ödenmiş)</span>;
-        } else if (invoice.isOverdue) {
-          return <span className="badge badge-error">Satış Faturası ({invoice.overdueDays} gün gecikmiş)</span>;
-        } else {
-          return <span className="badge badge-warning">Satış Faturası (Ödenmemiş)</span>;
-        }
-      }
-      
-      // Normal fatura kontrolü
-      if (docType.includes('fatura') || docType.includes('invoice')) {
-        if (invoice.status === 'paid') {
-          return <span className="badge badge-success">Fatura (Ödenmiş)</span>;
-        } else if (invoice.isOverdue) {
-          return <span className="badge badge-error">Fatura ({invoice.overdueDays} gün gecikmiş)</span>;
-        } else {
-          return <span className="badge badge-warning">Fatura (Ödenmemiş)</span>;
-        }
-      }
-    }
-    
-    // Genel durum kontrolü (belge türü belirsizse)
-    if (invoice.status === 'paid') {
-      return <span className="badge badge-success">Ödenmiş</span>;
-    } else if (invoice.isOverdue) {
-      return <span className="badge badge-error">{invoice.overdueDays} gün gecikmiş</span>;
-    } else {
-      return <span className="badge badge-warning">Ödenmemiş</span>;
-    }
   };
 
   if (loading) {
@@ -275,184 +121,38 @@ const CustomerDetail = () => {
                 </span>
               </div>
               {summary && (
-                <>
-                  <div className="flex justify-between items-center pt-2 border-t">
-                    <span className="text-gray-600 font-medium">Ödenmemiş Fatura Toplamı</span>
-                    <span className="text-lg font-bold text-red-600">{formatCurrency(summary.totalUnpaidAmount)}</span>
-                  </div>
-                  {summary.weightedDays > 0 && (
-                    <div className="flex justify-between items-center pt-2">
-                      <span className="text-gray-600 font-medium">Ağırlıklı Gün</span>
-                      <span className="text-lg font-bold text-orange-600">{summary.weightedDays} gün</span>
-                    </div>
-                  )}
-                </>
+                <div className="flex justify-between items-center pt-2 border-t">
+                  <span className="text-gray-600 font-medium">Ödenmemiş Fatura Toplamı</span>
+                  <span className="text-lg font-bold text-red-600">{formatCurrency(summary.totalUnpaidAmount)}</span>
+                </div>
               )}
             </div>
           </div>
         </div>
       </div>
 
-      {/* Fatura Özeti */}
-      {summary && (
-        <div className="card">
-          <div className="card-header">
-            <h3 className="card-title flex items-center">
-              <FileText className="h-5 w-5 mr-2" />
-              Fatura Özeti
-            </h3>
-          </div>
-          <div className="card-content">
-            <div className="grid grid-cols-1 md:grid-cols-4 gap-4">
-              <div className="bg-blue-50 p-4 rounded-lg">
-                <div className="flex items-center">
-                  <FileText className="h-6 w-6 text-blue-600 mr-2" />
-                  <div>
-                    <p className="text-sm text-gray-600">Toplam Fatura</p>
-                    <p className="text-lg font-semibold">{summary.totalInvoices}</p>
-                  </div>
-                </div>
-              </div>
-              <div className="bg-green-50 p-4 rounded-lg">
-                <div className="flex items-center">
-                  <CheckCircle className="h-6 w-6 text-green-600 mr-2" />
-                  <div>
-                    <p className="text-sm text-gray-600">Ödenmiş</p>
-                    <p className="text-lg font-semibold">{summary.paidInvoices}</p>
-                  </div>
-                </div>
-              </div>
-              <div className="bg-red-50 p-4 rounded-lg">
-                <div className="flex items-center">
-                  <AlertTriangle className="h-6 w-6 text-red-600 mr-2" />
-                  <div>
-                    <p className="text-sm text-gray-600">Ödenmemiş</p>
-                    <p className="text-lg font-semibold">{summary.unpaidInvoices}</p>
-                  </div>
-                </div>
-              </div>
-              <div className="bg-orange-50 p-4 rounded-lg">
-                <div className="flex items-center">
-                  <DollarSign className="h-6 w-6 text-orange-600 mr-2" />
-                  <div>
-                    <p className="text-sm text-gray-600">Toplam Borç</p>
-                    <p className="text-lg font-semibold">{formatCurrency(summary.totalUnpaidAmount)}</p>
-                  </div>
-                </div>
-              </div>
-            </div>
-          </div>
-        </div>
+      {/* Tabs */}
+      <div role="tablist" className="tabs tabs-boxed">
+        <a
+          role="tab"
+          className={`tab ${activeTab === 'invoices' ? 'tab-active' : ''}`}
+          onClick={() => setActiveTab('invoices')}
+        >
+          Faturalar
+        </a>
+        <a
+          role="tab"
+          className={`tab ${activeTab === 'payments' ? 'tab-active' : ''}`}
+          onClick={() => setActiveTab('payments')}
+        >
+          Ödemeler
+        </a>
+      </div>
+
+      {activeTab === 'invoices' && (
+        <InvoiceList customerId={id} onSummaryLoaded={setSummary} />
       )}
-
-      {/* Fatura Listesi */}
-      <div className="card">
-        <div className="card-header">
-          <h3 className="card-title flex items-center">
-            <FileText className="h-5 w-5 mr-2" />
-            Fatura Listesi
-          </h3>
-        </div>
-        <div className="card-content">
-          {loadingInvoices ? (
-            <div className="flex items-center justify-center h-32">
-              <div className="animate-spin rounded-full h-16 w-16 border-b-2 border-primary-600"></div>
-            </div>
-          ) : invoices.length > 0 ? (
-            <div className="overflow-x-auto">
-              <table className="table">
-                                 <thead>
-                   <tr>
-                     <th>Fatura Tarihi</th>
-                     <th>Vade Tarihi</th>
-                     <th>Geçen Gün</th>
-                     <th>Tutar</th>
-                     <th>Durum</th>
-                     <th>Son Ödeme</th>
-                     <th>Ödeme Yöntemi</th>
-                   </tr>
-                 </thead>
-                <tbody>
-                                     {invoices.map((invoice) => (
-                     <tr key={invoice.id}>
-                       <td>{formatDate(invoice.date)}</td>
-                       <td>{invoice.dueDate ? formatDate(invoice.dueDate) : '-'}</td>
-                       <td>{getElapsedDays(invoice)} gün</td>
-                       <td className="font-medium">{formatCurrency(invoice.amount)}</td>
-                       <td>{getStatusBadge(invoice)}</td>
-                       <td>
-                         {invoice.status === 'paid' && invoice.lastPaymentDate 
-                           ? formatDate(invoice.lastPaymentDate) 
-                           : '-'
-                         }
-                       </td>
-                       <td>
-                         {invoice.status === 'paid' && invoice.paymentMethod ? (
-                           <span className="badge badge-success">{invoice.paymentMethod}</span>
-                         ) : (
-                           '-'
-                         )}
-                       </td>
-                     </tr>
-                   ))}
-                </tbody>
-              </table>
-            </div>
-          ) : (
-            <div className="text-center py-8">
-              <FileText className="mx-auto h-8 w-8 text-gray-400 mb-2" />
-              <p className="text-gray-500">Bu müşterinin faturası bulunmuyor</p>
-            </div>
-          )}
-        </div>
-      </div>
-
-      {/* Ödeme Listesi */}
-      <div className="card">
-        <div className="card-header">
-          <h3 className="card-title flex items-center">
-            <DollarSign className="h-5 w-5 mr-2" />
-            Ödeme Listesi
-          </h3>
-        </div>
-        <div className="card-content">
-          {loadingPayments ? (
-            <div className="flex items-center justify-center h-32">
-              <div className="animate-spin rounded-full h-16 w-16 border-b-2 border-primary-600"></div>
-            </div>
-          ) : payments.length > 0 ? (
-            <div className="overflow-x-auto">
-              <table className="table">
-                <thead>
-                  <tr>
-                    <th>Ödeme Tarihi</th>
-                    <th>Tutar</th>
-                    <th>Ödeme Yöntemi</th>
-                    <th>Açıklama</th>
-                  </tr>
-                </thead>
-                <tbody>
-                  {payments.map((payment) => (
-                    <tr key={payment.id}>
-                      <td>{formatDate(payment.date)}</td>
-                      <td className="font-medium text-green-600">{formatCurrency(payment.amount)}</td>
-                      <td>
-                        <span className="badge badge-success">{payment.documentType}</span>
-                      </td>
-                      <td>{payment.description || '-'}</td>
-                    </tr>
-                  ))}
-                </tbody>
-              </table>
-            </div>
-          ) : (
-            <div className="text-center py-8">
-              <DollarSign className="mx-auto h-8 w-8 text-gray-400 mb-2" />
-              <p className="text-gray-500">Bu müşterinin ödemesi bulunmuyor</p>
-            </div>
-          )}
-        </div>
-      </div>
+      {activeTab === 'payments' && <PaymentList customerId={id} />}
     </div>
   );
 };
